+++ docs-website/config/navigation.js
@@ -146,6 +146,15 @@ const navigation = [
 		],
 	},
 	{
+		title: 'Cloud',
+		links: [
+			{
+				title: 'Deploy to WunderGraph Cloud',
+				href: '/docs/cloud/deployments',
+			},
+		],
+	},
+	{
 		title: 'Deployment',
 		links: [
 			{
@@ -174,8 +183,8 @@ const navigation = [
 				href: '/docs/examples/nextjs',
 			},
 			{
-				title: 'NextJS + SWR',
-				href: '/docs/examples/nextjs-swr',
+				title: 'Vite + SWR',
+				href: '/docs/examples/vite-swr',
 			},
 			{
 				title: 'Hooks',