+++ packages/swr/src/hooks.ts
@@ -61,7 +61,7 @@ export const createHooks = <Operations extends OperationsDefinition>(client: Cli
 	const useQuery: UseQueryHook<Operations> = (options) => {
 		const { operationName, liveQuery, enabled = true, input, ...swrConfig } = options;
 		const { onSuccess: onSuccessProp, onError: onErrorProp } = swrConfig;
-		const key = { operationName, input };
+		const key = liveQuery ? { operationName, input, liveQuery } : { operationName, input };
 		const _key = serialize(key);
 
 		const response = useSWR(enabled ? key : null, !liveQuery ? queryFetcher : null, swrConfig);
@@ -293,10 +293,11 @@ export const createHooks = <Operations extends OperationsDefinition>(client: Cli
 			operationName,
 			input,
 			subscribeOnce,
+			resetOnMount,
 			onSuccess: onSuccessProp,
 			onError: onErrorProp,
 		} = options;
-		const key = { operationName, input };
+		const key = { operationName, input, subscription: true };
 		const _key = serialize(key);
 
 		const { data, error } = useSWR(enabled ? key : null, null);
@@ -317,10 +318,11 @@ export const createHooks = <Operations extends OperationsDefinition>(client: Cli
 
 		const { isLoading, isSubscribed } = useSubscribeTo({
 			mutationKey: _key,
+			enabled,
 			operationName,
 			input,
 			subscribeOnce,
-			enabled,
+			resetOnMount,
 			onSuccess,
 			onError,
 		});